# jtech-tasklist-backend/src/main/java/br/com/jtech/tasklist/adapters/protocols/tasklist/UpdateTasklistRequest.java
package br.com.jtech.tasklist.adapters.protocols.tasklist;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class UpdateTasklistRequest {
    
    @NotBlank(message = "Title is required")
    @Size(min = 1, max = 255, message = "Title must be between 1 and 255 characters")
    private String title;
}